messaging/requirements.txt
# 📱 DEPENDENCIAS PARA SISTEMA DE MENSAJERÍA
# ==========================================

# Bot de Telegram
python-telegram-bot==20.7

# Bot de WhatsApp (Flask para webhooks)
Flask==3.0.0
requests==2.31.0

# Procesamiento de texto y fecha
python-dateutil==2.8.2

# Logging mejorado
colorlog==6.8.0

# Validación de datos
pydantic==2.5.2

# Variables de entorno
python-dotenv==1.0.0

# HTTP cliente robusto
httpx==0.25.2

# Procesamiento asíncrono
asyncio-mqtt==0.16.1

# Seguridad
cryptography==41.0.8
